Remove unnecessary casts to uint
More stragglers from the Species int->ushort rewrite and Form int->byte

+++ PKHeX.Core/Legality/Encounters/Generator/Specific/EncounterEggGenerator.cs
@@ -59,7 +59,7 @@ public static class EncounterEggGenerator
 
         var otherSplit = species;
         (species, form) = GetBaseSpecies(chain, 1);
-        if ((uint)species == otherSplit)
+        if (species == otherSplit)
             yield break;
 
         if (species <= max)
