+++ PKHeX.Core/Legality/LearnSource/Sources/LearnSource2GS.cs
@@ -41,7 +41,7 @@ public sealed class LearnSource2GS : ILearnSource<PersonalInfo2>, IEggSource
 
     public ReadOnlySpan<ushort> GetEggMoves(ushort species, byte form)
     {
-        if ((uint)species > MaxSpecies)
+        if (species > MaxSpecies)
             return ReadOnlySpan<ushort>.Empty;
         return EggMoves[species].Moves;
     }
