+++ PKHeX.Core/Saves/Substructures/Gen8/LA/Pokedex/PokedexSave8a.cs
@@ -32,7 +32,7 @@ public sealed class PokedexSave8a
     public static ushort GetDexIndex(PokedexType8a which, ushort species)
     {
         // Check species is valid
-        if ((uint)species > MaxSpeciesID)
+        if (species > MaxSpeciesID)
             return DexInvalid;
 
         // Check each form
